requirements.txt
Django==1.6.5
django-filter==0.7
django-registration==1.0
django-rest-auth==0.2.3
djangorestframework==2.3.14
djangorestframework-chain==0.1.3
